% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cria_playlist_artistas.R
\name{cria_playlist_artistas}
\alias{cria_playlist_artistas}
\title{Criar Playlist no Spotify com Top 10 musicas de cada artista}
\usage{
cria_playlist_artistas(
  artistas,
  nome,
  descricao = NULL,
  publica = TRUE,
  colaborativa = FALSE
)
}
\arguments{
\item{artistas}{vetor com nome dos artistas}

\item{nome}{um nome para a playlist que será criada}

\item{descricao}{uma descrição para a playlist que será criada. opcional}

\item{publica}{TRUE se quer que a playlist seja pública. FALSE se quer que a playlist seja privada. Valor default: TRUE}

\item{colaborativa}{TRUE se quer que a playlist seja colaborativa. FALSE caso contrário. Valor default: FALSE}
}
\value{
retorna o link da playlist no Spotify
}
\description{
Essa função cria uma playlist no Spotify, e adiciona 10 musicas mais famosas de cada banda especificada.
}
\examples{
# criar_playlist(data.frame(
# uri = c("spotify:track:3Wrjm47oTz2sjIgck11l5e",
# "spotify:track:02MWAaffLxlfxAUY7c5dvx",
# "spotify:track:0VjIjW4GlUZAMYd2vXMi3b",
# "spotify:track:37BZB0z9T8Xu7U3e65qxFy",
# "spotify:track:07MDkzWARZaLEdKxo6yArG")),
# nome_playlist = "Nome da Playlist",
# descricao = "descricao da playlist",
# publica = FALSE, colaborativa = FALSE)
}
