% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definir_ids.R
\name{definir_ids}
\alias{definir_ids}
\title{Define ID's}
\usage{
definir_ids(user_id, client_id, client_secret)
}
\arguments{
\item{user_id}{é o seu nome de usuário (id único) no Spotify. Você pode consultá-lo nesse link: https://www.spotify.com/br/account/overview/}

\item{client_id}{o client_id é um dos tokens de autenticação da API do Spotify. Você pode obtê-lo nesse link: https://developer.spotify.com/dashboard/applications}

\item{client_secret}{o client_secret é um dos tokens de autenticação da API do Spotify. Você pode obtê-lo nesse link: https://developer.spotify.com/dashboard/applications}
}
\value{
não retorna nada
}
\description{
Para acessar a API do Spotify, são necessários tokens de autenticação.
Essa função vai definir esses tokens/ID'S como variáveis globais, para que não
precisem ser especificados sempre que uma consulta for realizada.
}
\examples{
definir_ids("meu_nome_de_usuario", "xxxxxxxxxxxxxxxxxxxxx", "xxxxxxxxxxxxxxxxxxxxx")
}
