% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_artistas_seguidos.R
\name{obter_artistas_seguidos}
\alias{obter_artistas_seguidos}
\title{Obtém Seus Arittas Seguidos no Spotify}
\usage{
obter_artistas_seguidos()
}
\value{
uma tibble com todos os seus artistas seguidos, e as seguintes informações sobre eles: c("name", "id", "genres", "images", "popularity", "followers_total")
}
\description{
Essa função pega informações de todas as seus artistas seguidos no Spotify.
}
\examples{
# obter_artistas_seguidos()
}
