% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_dados_musicas.R
\name{obter_dados_musicas}
\alias{obter_dados_musicas}
\title{Obter Dados de Musicas}
\usage{
obter_dados_musicas(ids_musicas)
}
\arguments{
\item{ids_musicas}{um vetor com os ID's das músicas que deseja obter informação}
}
\value{
uma tibble com as seguintes informações sobre as músicas especificadas: c("track_id","explicit", "album", "album_id", "artist_id", "track_popularity", "track_name", "artist_name", "genres", "artist_popularity", "followers", "danceability", "energy", "key", "loudness", "mode", "speechiness", "acousticness", "instrumentalness", "liveness", "valence", "tempo", "uri", "track_href", "analysis_url", "duration_ms", "time_signature" )
}
\description{
Essa função obtem dados de musicas. Tais como:
}
\examples{
# obter_dados_musicas(c("3Wrjm47oTz2sjIgck11l5e",
# "02MWAaffLxlfxAUY7c5dvx", "0VjIjW4GlUZAMYd2vXMi3b",
# "37BZB0z9T8Xu7U3e65qxFy", "07MDkzWARZaLEdKxo6yArG"))
}
